Merged PR 254263: Add support for reporting memory statistics on types
This PR is adding a new optional field `TypesToReportStatisticsOn` to `StressUtilOptions`. It is interpreted as a regular expression and makes the stress library calculate inclusive and exclusive retained bytes of objects with types matching the regex. This calculation is done on the baseline and final dumps, logged, and reported in telemetry as four new properties:

**TypeStatsExclusiveRetainedBytes_Base**
**TypeStatsInclusiveRetainedBytes_Base**
**TypeStatsExclusiveRetainedBytes_Final**
**TypeStatsInclusiveRetainedBytes_Final**

If `TypesToReportStatisticsOn` is not specified, the library behaves the same as without this change, i.e. no functional changes.